webapp/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.20
django-bootstrap-v5==1.0.11
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
python-decouple==3.8
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
